+++ @ecomplus/storefront-components/src/js/TheProduct.js
@@ -4,8 +4,10 @@ import {
   i19close,
   i19days,
   i19discountOf,
+  // i19endsIn,
   i19freeShippingFrom,
   i19loadProductErrorMsg,
+  // i19offer,
   i19only,
   i19outOfStock,
   i19paymentOptions,
@@ -104,10 +106,12 @@ export default {
       type: String,
       default: 'col-12 col-md-6'
     },
+    hasPromotionTimer: Boolean,
     hasStickyBuyButton: {
       type: Boolean,
       default: true
     },
+    hasQuantitySelector: Boolean,
     canAddToCart: {
       type: Boolean,
       default: true
@@ -148,13 +152,15 @@ export default {
       selectedVariationId: null,
       currentGalleyImg: 1,
       isOnCart: false,
+      qntToBuy: 1,
       isStickyBuyVisible: false,
       isFavorite: false,
       hasClickedBuy: false,
       hasLoadError: false,
       paymentOptions: [],
       customizations: [],
-      kitItems: []
+      kitItems: [],
+      currentTimer: null
     }
   },
 
@@ -163,8 +169,16 @@ export default {
     i19close: () => i18n(i19close),
     i19days: () => i18n(i19days),
     i19discountOf: () => i18n(i19discountOf),
+    i19endsIn: () => i18n({
+      pt_br: 'Acaba em',
+      en_us: 'Ends in'
+    }),
     i19freeShippingFrom: () => i18n(i19freeShippingFrom),
     i19loadProductErrorMsg: () => i18n(i19loadProductErrorMsg),
+    i19offer: () => i18n({
+      pt_br: 'Oferta',
+      en_us: 'Offer'
+    }),
     i19only: () => i18n(i19only),
     i19outOfStock: () => i18n(i19outOfStock),
     i19paymentOptions: () => i18n(i19paymentOptions),
@@ -226,9 +240,19 @@ export default {
 
     discount () {
       const { body } = this
+      const priceValue = this.fixedPrice || getPrice(body)
       return checkOnPromotion(body)
-        ? Math.round(((body.base_price - getPrice(body)) * 100) / body.base_price)
+        ? Math.round(((body.base_price - priceValue) * 100) / body.base_price)
         : 0
+    },
+
+    isOnSale () {
+      const { body } = this
+      return this.hasPromotionTimer &&
+        checkOnPromotion(body) &&
+        body.price_effective_date &&
+        body.price_effective_date.end &&
+        Date.now() < new Date(body.price_effective_date.end).getTime()
     },
 
     ghostProductForPrices () {
@@ -380,7 +404,7 @@ export default {
       const { customizations } = this
       this.$emit('buy', { product, variationId, customizations })
       if (this.canAddToCart) {
-        ecomCart.addProduct({ ...product, customizations }, variationId)
+        ecomCart.addProduct({ ...product, customizations }, variationId, this.qntToBuy)
       }
       this.isOnCart = true
     },
@@ -538,6 +562,37 @@ export default {
         obs.observe()
       }
       setStickyBuyObserver()
+    }
+    if (this.isOnSale) {
+      const [currentIsoDay] = new Date().toISOString().split('T', 2)
+      const targetTime = new Date(this.body.price_effective_date.end)
+      const now = Date.now()
+      if (targetTime.getTime() > now) {
+        const dayMs = 24 * 60 * 60 * 1000
+        const formatTime = (number) => number < 10 ? `0${number}` : number
+        const getRemainingTime = (targetDate) => {
+          const distance = targetDate.getTime() - Date.now()
+          const days = Math.floor(distance / dayMs)
+          const hours = Math.floor((distance % dayMs) / (1000 * 60 * 60))
+          const minutes = Math.floor((distance % (1000 * 60 * 60)) / (1000 * 60))
+          const seconds = Math.floor((distance % (1000 * 60)) / 1000)
+          return (days > 0 ? `${formatTime(days)}:` : '') +
+            `${formatTime(hours)}:${formatTime(minutes)}:${formatTime(seconds)}`
+        }
+        const daysBetween = Math.floor((targetTime.getTime() - now) / dayMs)
+        const targetDate = daysBetween > 2
+          ? new Date(`${currentIsoDay}T${targetTime.toISOString().split('T')[1]}`)
+          : targetTime
+        this.currentTimer = setInterval(() => {
+          this.$refs.timer.innerHTML = getRemainingTime(targetDate)
+        }, 1000)
+      }
+    }
+  },
+
+  destroyed () {
+    if (this.currentTimer) {
+      clearInterval(this.currentTimer)
     }
   }
 }