+++ @ecomplus/widget-product/src/index.js
@@ -14,10 +14,11 @@ export default (options = {}, elId = 'product') => {
     const $dock = document.getElementById(`${elId}-dock`)
     const isSSR = Boolean($dock)
     const { storefront } = window
-    let getScopedSlots, body
+    let getScopedSlots, body, themeConfig
     if (storefront) {
       getScopedSlots = storefront.getScopedSlots
       body = storefront.context && storefront.context.body
+      themeConfig = storefront.theme && storefront.theme.product
     }
 
     let mounted
@@ -34,9 +35,17 @@ export default (options = {}, elId = 'product') => {
 
     const {
       buyText,
+      strHasQuantitySelector,
+      strHasPromotionTimer,
       lowQuantityToWarn,
       maxVariationOptionsBtns
     } = options
+
+    const strOptionToBool = (strOption, prop) => {
+      return strOption === '_'
+        ? Boolean(themeConfig && themeConfig[prop])
+        : strOption ? Boolean(strOption.trim()) : false
+    }
 
     new Vue({
       render: h => h(TheProduct, {
@@ -47,6 +56,8 @@ export default (options = {}, elId = 'product') => {
           ...options.props,
           product: isSSR && body && body.available && checkInStock(body) ? body : null,
           buyText,
+          hasQuantitySelector: strOptionToBool(strHasQuantitySelector, 'hasQuantitySelector'),
+          hasPromotionTimer: strOptionToBool(strHasPromotionTimer, 'hasPromotionTimer'),
           lowQuantityToWarn,
           maxVariationOptionsBtns,
           isSSR