+++ @ecomplus/widget-product-card/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.0.0-beta.114](https://github.com/ecomplus/storefront/compare/@ecomplus/widget-product-card@1.0.0-beta.113...@ecomplus/widget-product-card@1.0.0-beta.114) (2022-03-05)
+
+**Note:** Version bump only for package @ecomplus/widget-product-card
+
+
+
+
+
 # [1.0.0-beta.113](https://github.com/ecomplus/storefront/compare/@ecomplus/widget-product-card@1.0.0-beta.112...@ecomplus/widget-product-card@1.0.0-beta.113) (2022-02-16)
 
 **Note:** Version bump only for package @ecomplus/widget-product-card